General Update
Changed several things, mostly setup
+ utils-Class for utilities
+ utils::CheckForPHPVersion() to check if php-version is enough to run
+ reworked setup a bit

+++ admin/pages/user.overview.php
@@ -1,15 +1,18 @@
 <?php
 if($action == 'delete') {
-	
-		$sql = sql::factory();
-		$sql->setTable('user');
-		$sql->setWhere('id='.$id);
-		$sql->delete();
 		
+		if(dyn::get('user')->get('id') == $id)
+			echo message::danger(lang::get('user_self_deleted'));	
+		else {
+			$sql = sql::factory();
+			$sql->setTable('user');
+			$sql->setWhere('id='.$id);
+			$sql->delete();
+			
+			echo message::success(lang::get('user_deleted'));
+		}
+			
 		$action = '';
-		
-		echo message::success(lang::get('user_deleted'));
-		
 }
 
 if($action == 'add' || $action == 'edit') {
@@ -41,7 +44,7 @@ if($action == 'add' || $action == 'edit') {
 	$field->addValidator('email', lang::get('user_wrong_email'));
 	
 	if($form->get('password') != $form->sql->getValue('password')) {
-		$password = userLogin::hash($form->get('password'));
+		$password = userLogin::hash($form->get('password'), $form->get('salt'));
 	} else {
 		$password = $form->sql->getValue('password');
 	}
@@ -71,41 +74,28 @@ if($action == 'add' || $action == 'edit') {
 	}
 	
 	if($action == 'edit') {
+
 		$form->addHiddenField('id', $id);
-	}
+        $title = '"'.$form->get('firstname').' '.$form->get('name').'" '.lang::get('edit');
+
+	} else {
+
+        $title = lang::get('add');
+
+    }
 	
 	if($form->isSubmit()) {
 		
 		if($form->get('password') != $form->sql->getValue('password')) {
-			
-			$form->addPost('password', userLogin::hash($form->get('password')));
-			
+			$form->addPost('password', userLogin::hash($form->get('password'), $form->sql->getValue('salt')));
 		}
 		
 	}
+
+    $button = '<a href="'.url::backend('user', ['subpage'=>'overview']).'" class="btn btn-sm btn-default">'.lang::get('back').'</a>';
 	
 	?>
-	<div class="row">
-        <div class="col-lg-12">
-            <div class="panel panel-default">
-                <div class="panel-heading">
-                	<?php if($action == 'edit') { ?>
-                    	<h3 class="panel-title pull-left">"<?php echo $form->get('firstname')." ".$form->get('name'); ?>" <?php echo lang::get('edit'); ?></h3>
-                    <?php } else { ?>
-                    	<h3 class="panel-title pull-left"><?php echo lang::get('add'); ?></h3>
-                    <?php } ?>
-                    <div class="btn-group pull-right">
-						<a href="<?php echo url::backend('user', ['subpage'=>'overview']); ?>" class="btn btn-sm btn-default"><?php echo lang::get('back'); ?></a>
-					</div>
-					<div class="clearfix"></div>
-                </div>
-            	<div class="panel-body">
-					<?php echo $form->show(); ?>
-                </div>
-            </div>
-        </div>
-    </div>
-    
+	<div class="row"><?= bootstrap::panel($title, [$button], $form->show()); ?></div>
     <?php
 	
 }
@@ -129,7 +119,7 @@ if($action == '') {
 		$id = $table->get('id');
 			
 		$edit = '<a href="'.url::backend('user', ['subpage'=>'overview', 'action'=>'edit', 'id'=>$id]).'" class="btn btn-sm  btn-default fa fa-pencil-square-o"></a>';
-		$delete = '<a href="'.url::backend('user', ['subpage'=>'overview', 'action'=>'delete', 'id'=>$id]).'" class="btn btn-sm btn-danger fa fa-trash-o"></a>';
+		$delete = (dyn::get('user')->get('id') == $id) ? '' : '<a href="'.url::backend('user', ['subpage'=>'overview', 'action'=>'delete', 'id'=>$id]).'" class="btn btn-sm btn-danger fa fa-trash-o delete"></a>';
 		
 		$table->addRow()
 		->addCell($table->get('firstname')." ".$table->get('name'))
@@ -139,23 +129,11 @@ if($action == '') {
 		$table->next();	
 		
 	}
+
+    $button = '<a href="'.url::backend('user', ['subpage'=>'overview', 'action'=>'add']).'" class="btn btn-sm btn-default">'.lang::get('add').'</a>';
 	
 	?>
-    <div class="row">
-        <div class="col-lg-12">
-            <div class="panel panel-default">
-                <div class="panel-heading">
-                    <h3 class="panel-title pull-left"><?php echo lang::get('user'); ?></h3>
-					<div class="btn-group pull-right">
-						<a href="<?php echo url::backend('user', ['subpage'=>'overview', 'action'=>'add']); ?>" class="btn btn-sm btn-default"><?php echo lang::get('add'); ?></a>
-					</div>
-					<div class="clearfix"></div>
-                </div>
-                <?php echo $table->show(); ?>
-            </div>
-        </div>
-    </div>
-    
+    <div class="row"><?= bootstrap::panel(lang::get('user'), [$button], $table->show(), ['table' => true]) ?></div>
     <?php
 	
 }
